Add a new option in the workflow config shown in the Upload view to enable source language auto-detection in the Transcribe operator, and enable that option in the default workflow configuration.
Also, do not append the source language to the target language codes for WebToSRTCaptions and WebToVTTCaptions since 'auto' is an invalid code. In order to generate SRT and VTT for the source language, we will modify get_transcribe in MIE to put the source language into workflow metadata so downstream operators (such as the webcaptions operator) can get the valid source language code, whether it was auto detected or user specified.