Fix the pr-workflow in development
Background
==========

A github workflow runs when a PR is created targeting the development
branch. The workflow runs the build script and then deploys the stack
and runs automated tests against the deployed stack. The build fails
currently so the pr-workflow needs to be updated.

Change
======

Update the workflow to specify the required node.js version. The latest
runner was running node.js 18, which caused the build to fail. The
project currently needs to run in node.js 16.

Also updated deprecated actions, which were posting warnings and will
stop working soon.

See, https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12/