Add unit tests for lambda code written in Python
The solution had end-to-end tests but no unit tests so code coverage
scans reported 0% coverage. The solution includes code for three Lambda
functions under the `source` directory in addition to the website code
located in `source/website`. For now, we are not creating unit tests for
the web UI code so that leaves the Lambda functions, which are all
implemented in Python.

Added a `unit` directory under the `test` directory including a set of
unit tests for each of the three Lambda functions plus a runner script
to run the test cases. Also, updated `deployment/run-unit-tests.sh` to
run the new unit tests script so they will run in the build pipeline.

This change also includes some refactoring of the Lambda functions code
to resolve issues flagged by SonarQube scans.