Fix upload of WebVTT files
Background
==========

The **Upload Content** page contains an input box for uploading media
files and a text box to name a WebVTT file to use. It should allow the
WebVTT file to be uploaded and the text box should be automatically
populated with the name of the file. However, attempting to do so
resulted in an error stating the following:

> Unsupported file type: text/vtt

Change
======

This change fixes the issue. Now a user can select a WebVTT file. No
error is reported and the **Use Existing Subtitles** text box gets
automatically populated with the name of the WebVTT file.

Selecting "Upload and Run Workflow" causes the files to successfully
upload and the workflow begins to run.